+++ sdk/src/main/java/com/hedera/hashgraph/sdk/ManagedNode.java
@@ -1,19 +1,28 @@
 package com.hedera.hashgraph.sdk;
 
+import com.google.errorprone.annotations.Var;
 import io.grpc.ChannelCredentials;
+import io.grpc.ConnectivityState;
 import io.grpc.Grpc;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
 import io.grpc.TlsChannelCredentials;
 import io.grpc.inprocess.InProcessChannelBuilder;
+import java8.util.Objects;
 import org.threeten.bp.Duration;
 
 import javax.annotation.Nullable;
-import java.util.Objects;
+import java8.util.concurrent.CompletableFuture;
+
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.TimeUnit;
 
 abstract class ManagedNode<N extends ManagedNode<N, KeyT>, KeyT> implements Comparable<ManagedNode<N, KeyT>> {
+    private static final int GET_STATE_INTERVAL_MILLIS = 50;
+    private static final int GET_STATE_TIMEOUT_MILLIS = 10000;
+    private static final int GET_STATE_MAX_ATTEMPTS = GET_STATE_TIMEOUT_MILLIS / GET_STATE_INTERVAL_MILLIS;
+    private boolean hasConnected = false;
+
     protected final ExecutorService executor;
 
     /**
@@ -216,7 +225,7 @@ abstract class ManagedNode<N extends ManagedNode<N, KeyT>, KeyT> implements Comp
         if (channel != null) {
             return channel;
         }
-        
+
         ManagedChannelBuilder<?> channelBuilder;
 
         if (address.isInProcess()) {
@@ -228,11 +237,45 @@ abstract class ManagedNode<N extends ManagedNode<N, KeyT>, KeyT> implements Comp
         }
 
         channel = channelBuilder
+            .keepAliveTimeout(10, TimeUnit.SECONDS)
             .userAgent(getUserAgent())
             .executor(executor)
             .build();
 
         return channel;
+    }
+
+    boolean channelFailedToConnect() {
+        if (hasConnected) {
+            return false;
+        }
+        hasConnected = (getChannel().getState(true) == ConnectivityState.READY);
+        try {
+            for (@Var int i = 0; i < GET_STATE_MAX_ATTEMPTS && !hasConnected; i++) {
+                TimeUnit.MILLISECONDS.sleep(GET_STATE_INTERVAL_MILLIS);
+                hasConnected = (getChannel().getState(true) == ConnectivityState.READY);
+            }
+        } catch (InterruptedException e) {
+            throw new RuntimeException(e);
+        }
+        return !hasConnected;
+    }
+
+    private CompletableFuture<Boolean> channelFailedToConnectAsync(int i, ConnectivityState state) {
+        hasConnected = (state == ConnectivityState.READY);
+        if (i >= GET_STATE_MAX_ATTEMPTS || hasConnected) {
+            return CompletableFuture.completedFuture(!hasConnected);
+        }
+        return Delayer.delayFor(GET_STATE_INTERVAL_MILLIS, executor).thenCompose(ignored -> {
+            return channelFailedToConnectAsync(i + 1, getChannel().getState(true));
+        });
+    }
+
+    CompletableFuture<Boolean> channelFailedToConnectAsync() {
+        if (hasConnected) {
+            return CompletableFuture.completedFuture(false);
+        }
+        return channelFailedToConnectAsync(0, getChannel().getState(true));
     }
 
     /**